Add reminder about leading and trailing whitespaces in the strings.
String should not contain leading and trailing
whitespaces.
It is problematic to sort such strings.
Truncation of the string could be done on the server side.
After string is added and if it is displayed
somewhere on the page,
it could be performed a test to check the presence
of leading and trailing whitespaces.

With current demo it is not possible to do such test.